requirements.txt
django==3.0.2
python-telegram-bot==12.2.0
